Flatten CoverageExtension components folder
