--- packages/slate-editor/src/modules/editor-v4-coverage/components/CoverageElement/CoverageCard/index.ts
@@ -1,1 +1,0 @@
-export { CoverageCard } from './CoverageCard';
