--- packages/slate-editor/src/modules/editor-v4-coverage/components/CoverageElement/index.ts
@@ -1,1 +1,0 @@
-export { CoverageElement } from './CoverageElement';
