+++ packages/slate-editor/src/modules/editor-v4-coverage/components/FloatingCoverageMenu.module.scss
