--- packages/slate-editor/src/modules/editor-v4-coverage/components/FloatingCoverageMenu/index.ts
@@ -1,1 +1,0 @@
-export { FloatingCoverageMenu } from './FloatingCoverageMenu';
